refactor(beavernest-app): place readiness adapter in platform ring
