+++ apps/beavernest-app/lib/platform/web/readiness_client.dart
