+++ apps/beavernest-app/test/generated_contract_test.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:beavernest_app/api/readiness_client.dart';
+import 'package:beavernest_app/platform/web/readiness_client.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:http/http.dart' as http;
